cleanup: remove myroom website hosting
Removing the src files for My Room, formerly one of my weekend projects
that was not maintained. The frontend uses Node.Js packages with known
security vulnerabilities and the backend has been deactivated.

This commit adds a zip of the original src for posterity (in case I need
code samples in the future). Long live random projects.